+++ src/device/net/rtl8139.c
@@ -39,8 +39,8 @@ void rtl_receive(struct rtl8139 *rtl)
 			memcpy(sb->data, rtl->rx_buffer + rtl->rx_offset + 4, len - offset);
 			//uintptr_t off2 = len - offset;
 			memcpy(sb->data + (len - offset), rtl->rx_buffer, offset);
-			network_received(sb); 
-			
+			network_received(sb);
+
 		}else{
 
 			printf("receive error\n");
@@ -69,8 +69,8 @@ void rtl_handler(void *aux)
 	}
 
 	if(irq & 0x1)
-	{	
-		
+	{
+
 		rtl_receive(rtl);
 
 		rtl_outw(rtl, ISR, 0x1);
@@ -84,29 +84,29 @@ void rtl8139_enable_timer(struct rtl8139 *rtl)
 	rtl_outl(rtl, 0x48, 0); //set timer count to 0
 }
 char lpacket[] = {
-0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x00, 0xc0, 
-0x9f, 0x50, 0x32, 0xa0, 0x08, 0x06, 0x00, 0x01, 
-0x08, 0x00, 0x06, 0x04, 0x00, 0x01, 0x00, 0xc0, 
-0x9f, 0x50, 0x32, 0xa0, 0x80, 0x87, 0x9b, 0xcb, 
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x87, 
-0x9b, 0x58, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x00, 0xc0,
+0x9f, 0x50, 0x32, 0xa0, 0x08, 0x06, 0x00, 0x01,
+0x08, 0x00, 0x06, 0x04, 0x00, 0x01, 0x00, 0xc0,
+0x9f, 0x50, 0x32, 0xa0, 0x80, 0x87, 0x9b, 0xcb,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x87,
+0x9b, 0x58, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
 0x00, 0x00, 0x00, 0x00 };
 
 
 size_t rtl8139_send(struct network_dev *dev, uint8_t *_buf, size_t length)
 {
 	struct rtl8139 *rtl = dev->device;
-	
+
 	void* tx_buffer = (void *)(rtl->tx_buffers + 8192*rtl->tx_cur);
 	memset(tx_buffer, 0, (length <60) ? 60 : length);
 	memcpy(tx_buffer, _buf, length);
-		
+
 	if(length < 60)
 		length = 60;
-	
+
 	rtl_outl(rtl, 0x20 + rtl->tx_cur*4, (uint32_t)V2P(tx_buffer));
-	rtl_outl(rtl, 0x10 + rtl->tx_cur*4, length | (48 << 16)); 
+	rtl_outl(rtl, 0x10 + rtl->tx_cur*4, length | (48 << 16));
 	rtl->tx_cur++;
 	rtl->tx_cur %= 4;
 
@@ -127,11 +127,11 @@ void rtl8139_start(struct rtl8139 *rtl)
 	rtl->tx_cur = 0;
 	rtl_outb(rtl, 0x37, 0x10);
 	while((rtl_inb(rtl, 0x37) & 0x10) != 0);
-	
+
 	memset(rtl->rx_buffer, 0, (8192*8)+16+1500);
 	rtl_outl(rtl, 0x30,(uintptr_t)V2P(rtl->rx_buffer));
 	rtl_outb(rtl, 0x37, 0xc);
-	
+
 	for(int i=0; i < 4; i++)
 	{
 		rtl_outl(rtl, 0x20 + i*4, (uintptr_t)V2P(rtl->tx_buffers) + i*(8192 +16+1500));
@@ -154,8 +154,8 @@ void rtl8139_start(struct rtl8139 *rtl)
 			break;
 		}*/
 	}
-	
-	
+
+
 
 
 
@@ -183,11 +183,12 @@ struct network_dev * rtl8139_init()
 	if(rtl->pci != NULL)
 	{
 		rtl->pci_hdr = rtl->pci->header;
-		
+
 		rtl->io_base = pci_get_bar(rtl->pci, PCI_BAR_IO) & ~1;
-		
-		rtl->mem_base = (uint8_t *)(pci_get_bar(rtl->pci, PCI_BAR_MEM) & ~3);
-		
+
+		// XXX: This might be severely broken on 64bit machines
+		rtl->mem_base = (uint8_t *)(uintptr_t)(pci_get_bar(rtl->pci, PCI_BAR_MEM) & ~3);
+
 		rtl8139_start(rtl);
 		rtl8139_getmac(rtl,(char *)&device->mac);
 		device->send = rtl8139_send;
